diagchar_core: Fix pointer-to-int casts

+++ drivers/char/diag/diagchar_core.c
@@ -496,7 +496,11 @@ int diag_copy_remote(char __user * buf, size_t count, int *pret, int *pnum_data)
 
 		for (i = 0; i < diag_hsic[index].poolsize_hsic_write; i++) {
 			if (hsic_buf_tbl[i].length > 0) {
-				pr_debug("diag: HSIC copy to user, i: %d, buf: %x, len: %d\n", i, (unsigned int)hsic_buf_tbl[i].buf, hsic_buf_tbl[i].length);
+				pr_debug(
+					"diag: HSIC copy to user, i: %d, buf: %x, len: %d\n",
+					i,
+					(unsigned int) (uintptr_t) hsic_buf_tbl[i].buf,
+					hsic_buf_tbl[i].length);
 				num_data++;
 #if 0
 				/* Copy the negative token */
@@ -1885,7 +1889,8 @@ static ssize_t diagchar_write(struct file *file, const char __user * buf, size_t
 	/* This is to check if after HDLC encoding, we are still within the
 	   limits of aggregation buffer. If not, we write out the current buffer
 	   and start aggregation in a newly allocated buffer */
-	if ((unsigned int)enc.dest >= (unsigned int)(buf_hdlc + HDLC_OUT_BUF_SIZE)) {
+	if ((unsigned int) (uintptr_t) enc.dest >=
+	    (unsigned int) (uintptr_t)(buf_hdlc + HDLC_OUT_BUF_SIZE)) {
 		err = diag_device_write(buf_hdlc, APPS_DATA, NULL);
 		if (err) {
 			ret = -EIO;
@@ -1903,7 +1908,8 @@ static ssize_t diagchar_write(struct file *file, const char __user * buf, size_t
 		diag_hdlc_encode(&send, &enc);
 	}
 
-	driver->used = (uint32_t) enc.dest - (uint32_t) buf_hdlc;
+	driver->used =
+		(uint32_t)(uintptr_t) enc.dest - (uint32_t)(uintptr_t) buf_hdlc;
 	if (pkt_type == DATA_TYPE_RESPONSE) {
 		err = diag_device_write(buf_hdlc, APPS_DATA, NULL);
 		if (err) {
